+++ include/game/bullet.h
@@ -2,11 +2,14 @@
 #define BULLET_H
 
 #include "game.h"
+#include "weapon.h"
 class Game;
+class Weapon;
 
 class Bullet {
 private:
     Game* game;
+    Weapon* weapon;
     BulletData data;
     LevelPos pos;
     LevelVelocity velocity;
@@ -14,7 +17,7 @@ private:
     int reboundsLeft;
 
 public:
-    Bullet(Game* game, BulletData data, LevelPos start, LevelVelocity velocity);
+    Bullet(Game* game, Weapon* weapon, BulletData data, LevelPos start, LevelVelocity velocity);
     ~Bullet();
 
     void update();